feat(front): add auth guard and move app-wide features to core folder
- Add 'authGuard' for protected routes
- Update AppRoutingModule to use authGuard
- Add core folder for app-wide services and guards
- Move 'TokenService' to 'core/services' and 'AuthInterceptor' to 'core/interceptors'
