+++ front/src/app/core/guards/auth.guard.ts
@@ -1,0 +1,16 @@
+import { CanActivateFn, Router } from '@angular/router';
+import { inject } from '@angular/core';
+import { TokenService } from '../services/token.service';
+
+export const authGuard: CanActivateFn = (route, state) => {
+  const tokenService = inject(TokenService);
+  const router = inject(Router);
+
+  const token = tokenService.getToken();
+
+  if (token) {
+    return true; // with token access allowed
+  } else {
+    return router.createUrlTree(['/login']); // without a token redirect to login
+  }
+};
